requirements.txt
torch
numpy
higher
gym==0.12.1
imageio
mujoco_py
git+https://github.com/dennisl88/rand_param_envs.git@4d1529d61ca0d65ed4bd9207b108d4a4662a4da0#egg=rand_param_envs
hydra-core
